pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anima"
version = "0.1.0"
description = "Heartbeat-driven, distributed, self-evolving AI life system"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40.0",
    "psutil>=5.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
vector = ["chromadb>=0.5.0"]
watch = ["watchdog>=4.0.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
anima = "anima.main:main_entry"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["anima*"]
